+++ packages/uilib/src/lib/components/editable-text/index.ts
@@ -1,1 +1,4 @@
 export { default as EditableText } from "./editable-text.svelte"
+import type { API } from "./editable-text.svelte"
+export type EditableTextAPI = API
+// export { API as EditableTextAPI  } from "./editable-text.svelte"